Simple attempt to transcribe audio files.
Basically as of now this script can
- create an eaf file w/ same name as audio file
- add an annotation to the eaf file with the word spoken in that file.

Elpis documentation mentions it's okay for transcriptions to not be
at the individual sound level or even word level.
(https://elpis.readthedocs.io/en/latest/wiki/preparing-files.html)

But I think at the very least we should try to get the beginning of the
utterance correct.