Add German transcriptions.
Worth noting that while I was adding the transcriptions the program logged

INFO:root:Missing 1747 transcriptions of 3929 files

the 3929 number is how many files are listed in train.txt

I think that those are all supposed to be digit files??? Yet the transcription
of the file was allegedly missing from the tsv file.

I may investigate further.